more gcc fixes

+++ src/Kmers.cc
@@ -28,7 +28,7 @@ Kmer kmer_to_bits(const std::string & str) {
 std::set<jellyfish::mer_dna> kmers_in_string(const std::string & str) {
     auto k = jellyfish::mer_dna::k();
     set<jellyfish::mer_dna> s;
-    for (int i = 0; i <= str.size() - k; i++) {
+    for (size_t i = 0; i <= str.size() - k; i++) {
         s.insert(jellyfish::mer_dna(str.substr(i, k)));
     }
     return s;
